requirements.txt
torch>=1.8.1,<2
torchvision>=0.8.9,<2
tqdm>=2
scikit-learn>=1.0,<2
open_clip_torch>=0.2.1
pycocoevalcap
webdataset>=0.2.31
transformers
ovmsclient==2022.3.0
